Refactor MAPBUILDER
Instead of requiring instantiation of `generateMarkers`,
`categorizedMarkers` and `setMapCenter` for both Skatepark and User
pages, default behavior to skatepark and stub these methods/attributes
when building map for user. This pattern will be easier to extend when
we add in designers, and it also exposes `MAPBUILDER`'s API a little
more effectively. In the future when we want to allow users to
manipulate the map, it will be easier to add functionality following
this pattern.